LF-4769 Rename task services file
